function_data_conversion/requirements.txt
python-docx
jsons
google-cloud-storage
google-cloud-logging
fuzzywuzzy
python-Levenshtein
unidecode
pydantic<2.0
functions-framework
pytest